// src/app/models/TrackingHeader.ts
export class TrackingHeader
{
    CarrierCode: string;  
    DeliveryCity: string;
    DeliveredDateTime: string;
    DeliveryLocation:string;
    Description: string;
    DestinationLocation: string;
    DeviceId: string;
    EstDeliveryDate: string;
    InvoiceNo: string;
    Note1: string;
    OriginCity: string;
    OriginLocation: string;
    OriginZip: string;
    PackageType: string;
    Pieces: string;
    ServiceType: string; 
    SignedBy: string;
    ShipDate: string;
    SMMessage: string;
    SMTrackedTime: string;
    SourceData: string;
    SourceDate: string;
    Status:  string;
    Tracked: string;
    TrackingNo: string;
    UnitofMeasurement: string;
    Weight: string;  
}