// src/app/models/TrackingResult.ts
export class TrackingResult
{
  Archived: string;
  ATrDays: string;
  Billed_Weight: string;
  CarrierCode: string;
  Created: string;
  DateDelivered: any;
  DateGuaranteed: string;
  DateShipped: string;
  DeliveryLocation: string;
  Description: string;
  DestCity: string;
  DestCountry: string;
  DestState: string;
  DestZip: string;
  DeviceID: string;
  EnteredOn: string;
  ErrorMsg: string;
  EstDeliveryDate: any;
  GTrDays: string;
  InvoiceDate: string;
  InvoiceNo: string;
  LateExemptCode: string;
  OriginCity: string;
  OriginCountry: string;
  OriginState: string;
  OriginZip: string;
  RefundStatus: string;
  Residential: string;
  ResultCode: string;
  SaturdayDelivery: string;
  ScanActivityNote: string;
  ServiceType: string;
  ShipDateMethod: string;
  SignedBy: string;
  Status: string;
  Tracked: string;
  TrackingNo: string;
  Updated: string;
  Zone: string; 
}